add system table and procedures

+++ db/tests/conftest.py
@@ -1,3 +1,4 @@
+import json
 import os
 from uuid import uuid1
 
@@ -6,41 +7,66 @@ import pymysql
 import pytest
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope="session")
 def connection():
     connection = pymysql.connect(
-        host=os.getenv('MYSQL_HOST', '127.0.0.1'),
-        port=int(os.getenv('MYSQL_PORT', 3306)),
-        user='root',
-        password='testpassword',
-        database='spi_data',
-        binary_prefix=True)
+        host=os.getenv("MYSQL_HOST", "127.0.0.1"),
+        port=int(os.getenv("MYSQL_PORT", 3306)),
+        user="root",
+        password="testpassword",
+        database="spi_data",
+        binary_prefix=True,
+    )
     return connection
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope="session")
 def auth0_id():
-    return 'auth0|testuserid'
+    return "auth0|testuserid"
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope="session")
 def user_id():
     return str(uuid1())
 
 
-@pytest.fixture(scope='session')
-def standard_test_data(auth0_id, user_id, connection):
+@pytest.fixture(scope="session", params=[0, 1])
+def bad_user(request):
+    if request.param:
+        return "auth0|otheruser"
+    else:
+        return "invalid"
+
+
+@pytest.fixture(scope="session")
+def system_id():
+    return str(uuid1())
+
+
+@pytest.fixture(scope="session")
+def system_def():
+    return "A System", json.dumps({"version": "1", "other_parameters": []})
+
+
+@pytest.fixture(scope="session")
+def standard_test_data(auth0_id, user_id, connection, system_id, system_def):
     curs = connection.cursor()
+    otherid = str(uuid1())
+    curs.executemany(
+        "insert into users (id, auth0_id) values (uuid_to_bin(%s, 1), %s)",
+        [(user_id, auth0_id), (otherid, "auth0|otheruser")],
+    )
     curs.execute(
-        'insert into users (id, auth0_id) values (uuid_to_bin(%s, 1), %s)',
-        (user_id, auth0_id)
+        "insert into systems (id, user_id, name, definition) values "
+        "(uuid_to_bin(%s, 1), uuid_to_bin(%s, 1), %s, %s)",
+        (system_id, user_id, *system_def),
     )
     connection.commit()
     yield
-    curs.execute(
-        'delete from users where id = uuid_to_bin(%s, 1)',
-        user_id
+    curs.executemany(
+        "delete from users where id = uuid_to_bin(%s, 1)", (user_id, otherid)
     )
+    curs.execute("delete from systems where id = uuid_to_bin(%s, 1)", system_id)
     connection.commit()
